Extract logic to separate AreaStatsCard component
